Rename test suite topology_experimental_raft
