+++ test/topology_raft/test_tablets_removenode.py
