+++ test/topology_raft/test_topology_upgrade_not_stuck_after_recent_removal.py
