+++ test/topology_raft/test_zero_token_nodes_multidc.py
