+++ test/topology_raft/test_zero_token_nodes_no_replication.py
